W174d: btlFormation THREE MATCHes -- 475/520 (91.3%). 002CF070 340/352 nd178 -> 344/352 MATCH; 002CFEB0 352/272 nd272 -> 272/272 MATCH and window CLEARED; 002D0150 352/288 nd264 -> 280/288 MATCH and window CLEARED.
That completes the entire 002CE/002CF/002D0 family -- seven functions closed across three passes, every one first cleared of invented array zero-fills, u64 masks/widths and untyped pointer/call paths, then finished with sibling-grounded typed selection logic, explicit iterator/selected roles and retail-shaped positive guards.

Over-window 19 -> 17, LF-only preserved (0 CRLF, 15,156 LF), 46/46 pragma brackets and the retained opt_common_subs intact, the two intentional duplicate markers left exactly as they are. 002BAD60 stays at nd9 -- an argument-load-order probe regressed and was fully reverted.

+++ src/Event/Comu/comuFunction.c
@@ -1219,7 +1219,7 @@ int FUN_003beca0(int param_1)
 
   u32 uVar8;
 
-  long lVar9;
+  int lVar9;
 
   u8 *pbVar10;
